Whole-crate clippy + rustdoc cleanup: zero warnings under -D warnings
Drive both `cargo clippy --release --all-targets -- -D warnings` and
`RUSTDOCFLAGS=-D warnings cargo doc --no-deps --release` to zero
diagnostics across the entire workspace.

Clippy fixes (4):

* src/math.rs:235 — `i + 1 >= j + 1` simplified to `i >= j`
  (`clippy::int_plus_one`).
* src/rng/hash_drbg.rs:82-89 — drop the redundant `let drbg = Self {...}; drbg`
  (`clippy::let_and_return`).
* src/diehard/bitstream.rs:94 — `(bits_needed + 31) / 32` becomes
  `bits_needed.div_ceil(32)` (`clippy::manual_div_ceil`).
* src/research/webster_tavares.rs:254 — drop `as u64` no-op cast
  (`clippy::unnecessary_cast`).

Rustdoc fixes (13 broken intra-doc links, 3 unclosed HTML tags) — all
mathematical / shell notation that rustdoc was trying to resolve as
intra-doc links.  Wrap `[k]`, `[m]`, `[1600]`, etc. in backticks or
escape the brackets, and wrap `<name>`/`<count>` synopses in fenced
code blocks.  Files touched:

* src/bin/dump_rng.rs, src/bin/pilot_rng.rs
* src/diehard/birthday_spacings.rs
* src/dieharder/dct.rs, src/dieharder/minimum_distance_nd.rs
* src/rng/spongebob.rs

Verification:
* cargo clippy --release --all-targets -- -D warnings  → 0 warnings
* RUSTDOCFLAGS=-D warnings cargo doc --no-deps --release → 0 warnings
* cargo build --release  → 0 warnings
* cargo test --release  → 101 lib + 3 dump_rng unit + 7 dump_rng
  integration = 111 passing, 0 failing.

+++ src/research/webster_tavares.rs
@@ -251,7 +251,7 @@ mod tests {
     #[test]
     fn parity_broadcast_is_maximally_correlated() {
         let report = evaluate_u64(4, 4, 16, |x| {
-            let g = ((x & 1) & ((x >> 1) & 1)) as u64;
+            let g = (x & 1) & ((x >> 1) & 1);
             g | (g << 1)
         });
         for i in 0..2 {
